% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_data_1D}
\alias{ex_data_1D}
\title{Daily CO2 Efflux Measurements within a Temperate Forest}
\format{
\subsection{\code{ex_data_1D}}{

A data frame with 365 rows and 3 columns:
\describe{
\item{Time}{The day of the year}
\item{Temp}{The temperature}
\item{CO2}{The carbon dioxide efflux}
}
}
}
\source{
\url{https://doi.org/10.1007/s11104-017-3506-4}
}
\usage{
ex_data_1D
}
\description{
A dataset containing daily CO2 efflux levels as a dependent variable and
temperature as an independent variables within a temperate forest
for a full year.
}
\examples{
data(ex_data_1D)
}
\keyword{datasets}
