% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_data_2D}
\alias{ex_data_2D}
\title{Spatial Distribution of Soil CO2 Efflux for CONUS}
\format{
\subsection{\code{ex_data_2D}}{

A data frame with 903 rows and 4 columns:
\describe{
\item{X}{The longitude}
\item{Y}{The latitude}
\item{rTemp}{The temperature}
\item{rRS}{The carbon dioxide efflux}
}
}
}
\source{
\url{https://doi.org/10.1111/gcb.15666}
}
\usage{
ex_data_2D
}
\description{
A dataset containing spatially distributed CO2 efflux levels as a dependent
variable and temperature as an independent variables within CONUS.
}
\examples{
data(ex_data_2D)
}
\keyword{datasets}
